Refactor toolbar creation in AwsToolkitExplorerToolWindow & simplify utility function
