+++ plugins/toolkit/jetbrains-core/src-233/software/aws/toolkits/jetbrains/core/explorer/ConnectionActionToolbarBuilder.kt
@@ -7,28 +7,13 @@ import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.ActionPlaces
 import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.util.ui.components.BorderLayoutPanel
+import javax.swing.JComponent
 
 object ConnectionActionToolbarBuilder {
-    fun createToolbar(toolWindow: AwsToolkitExplorerToolWindow, group: DefaultActionGroup): BorderLayoutPanel {
-        return BorderLayoutPanel().apply {
-            addToCenter(
-                ActionManager.getInstance().createActionToolbar(ActionPlaces.TOOLBAR, group, true).apply {
-                    layoutPolicy = ActionToolbar.NOWRAP_LAYOUT_POLICY
-                    setTargetComponent(toolWindow)
-                }.component
-            )
-
-            val actionManager = ActionManager.getInstance()
-            val rightActionGroup = DefaultActionGroup(
-                actionManager.getAction("aws.toolkit.toolwindow.credentials.rightGroup.more"),
-                actionManager.getAction("aws.toolkit.toolwindow.credentials.rightGroup.help")
-            )
-
-            addToRight(
-                ActionManager.getInstance().createActionToolbar(ActionPlaces.TOOLBAR, rightActionGroup, true).apply {
-                    setTargetComponent(toolWindow.component)
-                }.component
-            )
-        }
+    fun createToolbar(targetComponent: JComponent, group: DefaultActionGroup): JComponent {
+        return ActionManager.getInstance().createActionToolbar(ActionPlaces.TOOLBAR, group, true).apply {
+            layoutPolicy = ActionToolbar.NOWRAP_LAYOUT_POLICY
+            setTargetComponent(targetComponent)
+        }.component
     }
 }
