+++ plugins/toolkit/jetbrains-core/src/software/aws/toolkits/jetbrains/core/explorer/AwsToolkitExplorerToolWindow.kt
@@ -6,7 +6,6 @@ package software.aws.toolkits.jetbrains.core.explorer
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.ActionPlaces
-import com.intellij.openapi.actionSystem.ActionToolbar
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.DefaultActionGroup
@@ -22,7 +21,6 @@ import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.ui.components.JBTabbedPane
 import com.intellij.util.ui.components.BorderLayoutPanel
 import software.aws.toolkits.jetbrains.core.credentials.CredsComboBoxActionGroup
-import software.aws.toolkits.jetbrains.core.explorer.ConnectionActionToolbarBuilder
 import software.aws.toolkits.jetbrains.core.explorer.cwqTab.CodewhispererQToolWindow
 import software.aws.toolkits.jetbrains.core.explorer.cwqTab.isQInstalled
 import software.aws.toolkits.jetbrains.core.explorer.devToolsTab.DevToolsToolWindow
@@ -64,8 +62,23 @@ class AwsToolkitExplorerToolWindow(
             val content = BorderLayoutPanel()
             setContent(content)
             val group = CredsComboBoxActionGroup(project)
+            val toolWindow = this
 
-            toolbar = ConnectionActionToolbarBuilder.createToolbar(this,group)
+            toolbar = BorderLayoutPanel().apply {
+                addToCenter(ConnectionActionToolbarBuilder.createToolbar(toolWindow, group))
+                val actionManager = ActionManager.getInstance()
+                val rightActionGroup = DefaultActionGroup(
+                    actionManager.getAction("aws.toolkit.toolwindow.credentials.rightGroup.more"),
+                    actionManager.getAction("aws.toolkit.toolwindow.credentials.rightGroup.help")
+                )
+
+                addToRight(
+                    ActionManager.getInstance().createActionToolbar(ActionPlaces.TOOLBAR, rightActionGroup, true).apply {
+                        // revisit if these actions need the tool window as a data provider
+                        setTargetComponent(component)
+                    }.component
+                )
+            }
 
             // main content
             tabComponents.forEach { name, contentProvider ->
